refactor(scale-feedback): rework the package's placement and public surface
