+++ packages/react/package.json
@@ -66,7 +66,7 @@
     "@seed-design/react-pull-to-refresh": "^2.1.0",
     "@seed-design/react-quantity-picker": "^1.0.0",
     "@seed-design/react-radio-group": "^2.0.1",
-    "@seed-design/react-scale-feedback": "workspace:*",
+    "@seed-design/react-scale-feedback": "^0.0.0",
     "@seed-design/react-segmented-control": "^2.0.2",
     "@seed-design/react-select": "^1.0.1",
     "@seed-design/react-side-navigation": "^1.0.1",
