+++ packages/utils/react-scale-feedback/LICENSE
