+++ packages/utils/react-scale-feedback/NOTICE
