+++ packages/utils/react-scale-feedback/tsconfig.json
